+++ SignNow.Net.Test/UnitTests/Services/UserServiceTest.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -7,6 +8,7 @@ using SignNow.Net.Model.Requests;
 using SignNow.Net.Service;
 using SignNow.Net.Test;
 using SignNow.Net.Test.Constants;
+using SignNow.Net.Test.FakeModels;
 
 namespace UnitTests
 {
@@ -128,6 +130,77 @@ namespace UnitTests
 #endif
             StringAssert.Contains(exception.Message, expectedMessage);
             Assert.AreEqual("unverified_email", exception.InnerExceptions[0].Message);
+        }
+
+        [TestMethod]
+        public void CreatesEmbeddedSigningInvite()
+        {
+            var mockResponse = @"
+                {
+                  ""data"": [
+                    {
+                      ""id"": ""139f1a566dc64946ab5fe73883811efd8136deb7"",
+                      ""email"": ""test.mail@gmail.com"",
+                      ""role_id"": ""c6fa0bf1f4e2045c7fbcf6335a26bef94b7199b7"",
+                      ""order"": 1,
+                      ""status"": ""pending""
+                    }
+                  ]
+                }";
+
+            var document = new SignNowDocumentFaker()
+                .FinishWith((f, o) =>
+                {
+                    o.Roles = new RoleFaker().Generate(1);
+                })
+                .Generate();
+            var embedded = new EmbeddedSigningInvite(document);
+            embedded.AddEmbeddedSigningInvite(
+                new EmbeddedInviteFaker()
+                    .FinishWith((f, o) =>
+                    {
+                        o.RoleId = document.Roles.First().Id;
+                    })
+                    .Generate());
+
+            var userService = new UserService(ApiBaseUrl, new Token(), SignNowClientMock(mockResponse));
+            var embeddedInvite = userService.CreateInviteAsync(document.Id, embedded).Result;
+
+            Assert.IsNotNull(embeddedInvite);
+            Assert.AreEqual("139f1a566dc64946ab5fe73883811efd8136deb7", embeddedInvite.InviteData[0].Id);
+            Assert.AreEqual("Pending", embeddedInvite.InviteData[0].Status.ToString());
+        }
+
+        [TestMethod]
+        public void GenerateEmbeddedSigningLink()
+        {
+            var mockResponse = @"
+            {
+              ""data"": {
+                ""link"": ""https://app-eval.signnow.com/webapp/document/9dbdffc9c5af49809d4dfdf613e9835b50a9582f?access_token=f108a715acd9272b7b25a3ecd2bc06962148de40b2ecdd08c05fbf41994b98b6&route=fieldinvite""
+              }
+            }";
+
+            var document = new SignNowDocumentFaker()
+                .FinishWith((f, o) =>
+                {
+                    o.Roles = new RoleFaker().Generate(1);
+                })
+                .Generate();
+
+            var options = new CreateEmbedLinkOptions
+            {
+                FieldInvite = new FieldInviteFaker().Generate(),
+                AuthMethod = EmbeddedAuthType.None,
+                LinkExpiration = 30
+            };
+
+            var userService = new UserService(ApiBaseUrl, new Token(), SignNowClientMock(mockResponse));
+            var embeddedLink = userService.GenerateEmbeddedInviteLinkAsync(document.Id, options).Result;
+            var link =
+                "https://app-eval.signnow.com/webapp/document/9dbdffc9c5af49809d4dfdf613e9835b50a9582f?access_token=f108a715acd9272b7b25a3ecd2bc06962148de40b2ecdd08c05fbf41994b98b6&route=fieldinvite";
+
+            Assert.AreEqual(link, embeddedLink.Link.AbsoluteUri);
         }
     }
 }